refactor: split MediaServiceImpl god class into focused services
- MediaQueryService: read operations (search, getDetails, getContent, getTags, etc.)
- MediaUploadService: upload/register + validateFileSignature
- MediaTransferService: copy/move/delete/update operations
- MediaServiceImpl: thin facade delegating to sub-services (1081→161 lines)
- Constructor deps reduced from 16 to 3 (queryService, uploadService, transferService)
- IMediaService interface preserved for backward compatibility
